backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
redis==6.4.0
celery==5.5.3
email-validator==2.2.0

